pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.9",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-sqlalchemy>=3.1.1",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
